msm: ais: reqmgr: Avoid freeing subdev twice
The 'l_device' pointer in __cam_req_mgr_destroy_subdev is
set to NULL after freeing but this is done on a
local copy of the variable in stack. This results in
double-free when this function is called again. To avoid
this, pass 'l_device' pointer by reference and assign it
to NULL after freeing.

Change-Id: Iebba12cc491d0a9ddfff612d72c3facbdc9187d5
Signed-off-by: E V Ravi <quic_evenka@quicinc.com>
(cherry picked from commit a9636c976a84f9aa460e45ff9ba0678ccd4c6a23)