irqchip: qcom: pdc: Don't overwrite IRQ domain name on init
The IRQ subsystem allocates a domain name dynamically on its own. By
overwriting it, there is not only a memory leak created, but also a
kfree() on a non-allocated pointer when the IRQ is freed. We can remedy
all of this by simply not overwriting the IRQ domain name.

Change-Id: I49a211de7db0020e0d4e91c1e0a0be8474a01441
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: atndko <z1281552865@gmail.com>
(cherry picked from commit 12c3d25b04011004b0bf7a70e50853d064470e9d)